requirements.txt
streamlit>=1.36
ultralytics==8.3.0
torch==2.2.2
torchvision==0.17.2
opencv-python-headless==4.9.0.80
pillow>=10.0.0
numpy>=1.24
huggingface_hub>=0.23.0
